+++ common/src/main/java/com/teamresourceful/resourcefulconfig/api/annotations/Config.java
@@ -1,6 +1,7 @@
 package com.teamresourceful.resourcefulconfig.api.annotations;
 
 import org.intellij.lang.annotations.Pattern;
+import org.jetbrains.annotations.Range;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -13,6 +14,9 @@ public @interface Config {
 
     @Pattern("^[a-z0-9_/-]+$")
     String value();
+
+    @Range(from = 0, to = Integer.MAX_VALUE)
+    int version() default 0;
 
     Class<?>[] categories() default {};
 }