+++ common/src/main/java/com/teamresourceful/resourcefulconfig/api/types/ResourcefulConfig.java
@@ -1,13 +1,19 @@
 package com.teamresourceful.resourcefulconfig.api.types;
 
+import com.teamresourceful.resourcefulconfig.api.patching.ConfigPatchEvent;
 import com.teamresourceful.resourcefulconfig.api.types.entries.ResourcefulConfigEntry;
 import com.teamresourceful.resourcefulconfig.api.types.info.ResourcefulConfigInfo;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.function.Consumer;
 
 public interface ResourcefulConfig {
+
+    default int version() {
+        return 0;
+    }
 
     @NotNull
     LinkedHashMap<String, ResourcefulConfigEntry> entries();
@@ -25,7 +31,7 @@ public interface ResourcefulConfig {
 
     void save();
 
-    void load();
+    void load(Consumer<ConfigPatchEvent> handler);
 
     boolean hasFile();