+++ common/src/main/java/com/teamresourceful/resourcefulconfig/client/components/header/HeaderControlsWidget.java
@@ -9,6 +9,7 @@ import net.minecraft.Optionull;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.StringWidget;
+import net.minecraft.client.gui.layouts.LayoutSettings;
 import net.minecraft.client.gui.layouts.LinearLayout;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
@@ -21,7 +22,7 @@ public class HeaderControlsWidget extends ContainerWidget {
     public HeaderControlsWidget(@Nullable String filename, int width) {
         super(0, 0, width, 0);
 
-        this.layout = LinearLayout.horizontal().spacing(UIConstants.PAGE_PADDING);
+        this.layout = LinearLayout.vertical().spacing(5);
 
         Screen screen = Minecraft.getInstance().screen;
         boolean willClose = screen instanceof ConfigScreen config && config.getParent() == null;
@@ -35,10 +36,12 @@ public class HeaderControlsWidget extends ContainerWidget {
 
         this.layout.addChild(
                 new StringWidget(
-                        this.width - 16 - UIConstants.PAGE_PADDING * 3, 16,
-                        Component.literal(Optionull.mapOrDefault(filename, name -> name + ".jsonc", "")).withColor(UIConstants.TEXT_PARAGRAPH), Minecraft.getInstance().font
-                ).alignRight(),
-                settings -> settings.alignVerticallyMiddle().alignHorizontallyRight()
+                        this.width, 10,
+                        Component.literal(Optionull.mapOrDefault(filename, name -> name + ".jsonc", ""))
+                                .withColor(UIConstants.TEXT_PARAGRAPH),
+                        Minecraft.getInstance().font
+                ).alignLeft(),
+                LayoutSettings::alignVerticallyMiddle
         );
 
         this.layout.arrangeElements();